build: migrate to jOOQ official codegen plugin and fix Docker networking

+++ task-scanner/src/main/java/ru/npepub/taskscanner/dto/SprintTaskInfo.java
@@ -1,4 +1,4 @@
 package ru.npepub.taskscanner.dto;
 
-public record SprintTaskInfo(long sprintNum, long taskNum) {
+public record SprintTaskInfo(int sprintNum, int taskNum) {
 }
